fix: address review comments (#18)
